View holders for Post adapter is now not inner classes
